removed duplicate breadcrumb
