+++ blocks/breadcrumbs/breadcrumbs-create.js
@@ -3,12 +3,9 @@ import {
 } from '../../scripts/dom-helpers.js';
 import ffetch from '../../scripts/ffetch.js';
 import { loadCSS } from '../../scripts/lib-franklin.js';
+import customBreadcrumbs, { customResourcesBreadcrumb } from './customBreadcrumbs.js';
 
 const customResourceTypes = ['Videos and Webinars', 'Application Note', 'Cell Counter', 'Interactive Demo'];
-const customResourcesBreadcrumb = {
-  name: 'Resources',
-  url_path: '/search-results',
-};
 
 function prependSlash(path) {
   return path.startsWith('/') ? path : `/${path}`;
@@ -18,130 +15,6 @@ function skipParts(pathSplit) {
   const partsToSkip = ['en', 'assets', 'br', 'img', 'citations', 'dd', 'tutorials-videos', 'bpd', 'cns', 'flipr', 'contaminants', 'enzyme'];
   return pathSplit.filter((item) => !partsToSkip.includes(item));
 }
-
-const customBreadcrumbs = {
-  'app-note': {
-    name: 'App Note',
-    url_path:
-      'https://www.moleculardevices.com/search-results#t=All&sort=relevancy&f:@md_contenttype=%5BApplication%20Note%5D',
-  },
-  ebook: {
-    name: 'EBook',
-    url_path:
-      'https://www.moleculardevices.com/search-results#t=All&sort=relevancy&f:@md_contenttype=%5BeBook%5D',
-  },
-  'lab-notes': {
-    name: 'Lab Notes',
-    url_path: '/lab-notes',
-  },
-  '/lab-notes/general': {
-    name: 'General',
-    url_path: '/lab-notes/blog#General',
-  },
-  '/lab-notes/clone-screening': {
-    name: 'Clone Screening',
-    url_path: '/lab-notes/blog#Clone-Screening',
-  },
-  '/lab-notes/cellular-imaging-systems': {
-    name: 'Cellular Imaging Systems',
-    url_path: '/lab-notes/blog#Cellular-Imaging-Systems',
-  },
-  '/lab-notes/microplate-readers': {
-    name: 'Microplate Readers',
-    url_path: '/lab-notes/blog#Microplate-Readers',
-  },
-  '/lab-notes/axon-patch-clamp': {
-    name: 'Axon Patch-Clamp',
-    url_path: '/lab-notes/blog#Axon-Patch-Clamp',
-  },
-  '/lab-notes/3d-biology': {
-    name: '3D Biology',
-    url_path: '/lab-notes/blog#3D-Biology',
-  },
-  'service-support': {
-    name: 'Service and Support',
-    url_path: '/service-support',
-  },
-  technology: {
-    name: 'Technology and Innovation',
-    url_path: '/technology',
-  },
-  'accessories-consumables': {
-    name: 'Accessories and Consumables',
-    url_path: '/products/accessories-consumables',
-  },
-  'customer-breakthrough': {
-    name: 'Customer Breakthrough',
-    url_path: '/customer-breakthroughs',
-  },
-  'acquisition-and-analysis-software': {
-    name: 'Acquisition and Analysis Software',
-  },
-  'igg-quantification-assays': {
-    name: 'IgG quantitation',
-  },
-  'dna-quantitation': {
-    name: 'DNA Quantitation',
-  },
-  elisa: {
-    name: 'ELISA',
-  },
-  'cell-viability': {
-    name: 'Cell Viability',
-  },
-  cardiotox: {
-    name: 'Cardiotox',
-  },
-  gpcrs: {
-    name: 'GPCR',
-  },
-  'ion-channel': {
-    name: 'Ion Channel',
-  },
-  'reporter-gene-assays': {
-    name: 'Reporter Gene',
-  },
-  'western-blot': {
-    name: 'Western Blot',
-  },
-  transporters: {
-    name: 'Transporter',
-  },
-  'mammalian-screening': {
-    name: 'Mammalian Screening',
-    url_path: '/products/clone-screening/mammalian-screening',
-  },
-  'microbial-screening': {
-    name: 'Microbial Screening',
-    url_path: '/products/clone-screening/microbial-screening',
-  },
-  stacker: {
-    name: 'Stacker',
-  },
-  'single-cell-isolation': {
-    name: 'Single-Cell Isolation',
-  },
-  digitizers: {
-    name: 'Digitizers',
-  },
-  amplifiers: {
-    name: 'Amplifiers',
-  },
-  resources: customResourcesBreadcrumb,
-  events: {
-    name: 'Events',
-    url_path: '/events',
-  },
-  newsletters: {
-    name: 'Newsletters',
-    url_path: '/newsletters',
-  },
-  brochures: customResourcesBreadcrumb,
-  'for-whats-next': {
-    name: 'For What\'s Next',
-    url_path: '/for-whats-next',
-  },
-};
 
 function getCustomUrl(path, part) {
   if (customBreadcrumbs[part]) {
@@ -156,22 +29,12 @@ function getCustomUrl(path, part) {
 }
 
 function getName(pageIndex, path, part, current) {
-  if (customBreadcrumbs[part]) {
-    return customBreadcrumbs[part].name;
-  }
-
-  if (customBreadcrumbs[path]) {
-    return customBreadcrumbs[path].name;
-  }
+  if (customBreadcrumbs[part]) return customBreadcrumbs[part].name;
+  if (customBreadcrumbs[path]) return customBreadcrumbs[path].name;
 
   const pg = pageIndex.find((page) => page.path === path);
-  if (pg && pg.h1 && pg.h1 !== '0') {
-    return pg.h1;
-  }
-
-  if (pg && pg.title && pg.title !== '0') {
-    return pg.title;
-  }
+  if (pg && pg.h1 && pg.h1 !== '0') return pg.h1;
+  if (pg && pg.title && pg.title !== '0') return pg.title;
 
   if (current) {
     const headingElement = document.querySelector('main h1');
@@ -196,13 +59,12 @@ export default async function createBreadcrumbs(container) {
 
   const pageIndex = await ffetch('/query-index.json').all();
   const pg = pageIndex.find((page) => page.path === path);
+
   // default Home breadcrumb
   const breadcrumbs = [
-    {
-      name: 'Home',
-      url_path: '/',
-    },
+    { name: 'Home', url_path: '/' },
   ];
+
   // custom resource types restricting breadcrumb to Home > Resources
   if (pg && customResourceTypes.includes(pg.type)) {
     breadcrumbs.push(customResourcesBreadcrumb);
@@ -211,6 +73,7 @@ export default async function createBreadcrumbs(container) {
     if (pg && pg.type === 'Customer Breakthrough') {
       breadcrumbs.push(customResourcesBreadcrumb);
     }
+
     // resolve rest of the path
     const urlForIndex = (index) => prependSlash(pathSplit.slice(1, index + 2).join('/'));
     breadcrumbs.push(
@@ -224,6 +87,8 @@ export default async function createBreadcrumbs(container) {
       { name: getName(pageIndex, path, pathSplit[pathSplit.length - 1], true) },
     );
   }
+
+  // render to ol list
   const ol = container.querySelector('ol');
   ol.setAttribute('itemscope', '');
   ol.setAttribute('itemtype', 'http://schema.org/BreadcrumbList');
