+++ blocks/breadcrumbs/customBreadcrumbs.js
@@ -1,0 +1,130 @@
+export const customResourcesBreadcrumb = {
+  name: 'Resources',
+  url_path: '/search-results',
+};
+
+const customBreadcrumbs = {
+  'app-note': {
+    name: 'App Note',
+    url_path:
+      'https://www.moleculardevices.com/search-results#t=All&sort=relevancy&f:@md_contenttype=%5BApplication%20Note%5D',
+  },
+  ebook: {
+    name: 'EBook',
+    url_path:
+      'https://www.moleculardevices.com/search-results#t=All&sort=relevancy&f:@md_contenttype=%5BeBook%5D',
+  },
+  'lab-notes': {
+    name: 'Lab Notes',
+    url_path: '/lab-notes',
+  },
+  '/lab-notes/general': {
+    name: 'General',
+    url_path: '/lab-notes/blog#General',
+  },
+  '/lab-notes/clone-screening': {
+    name: 'Clone Screening',
+    url_path: '/lab-notes/blog#Clone-Screening',
+  },
+  '/lab-notes/cellular-imaging-systems': {
+    name: 'Cellular Imaging Systems',
+    url_path: '/lab-notes/blog#Cellular-Imaging-Systems',
+  },
+  '/lab-notes/microplate-readers': {
+    name: 'Microplate Readers',
+    url_path: '/lab-notes/blog#Microplate-Readers',
+  },
+  '/lab-notes/axon-patch-clamp': {
+    name: 'Axon Patch-Clamp',
+    url_path: '/lab-notes/blog#Axon-Patch-Clamp',
+  },
+  '/lab-notes/3d-biology': {
+    name: '3D Biology',
+    url_path: '/lab-notes/blog#3D-Biology',
+  },
+  'service-support': {
+    name: 'Service and Support',
+    url_path: '/service-support',
+  },
+  technology: {
+    name: 'Technology and Innovation',
+    url_path: '/technology',
+  },
+  'accessories-consumables': {
+    name: 'Accessories and Consumables',
+    url_path: '/products/accessories-consumables',
+  },
+  'customer-breakthrough': {
+    name: 'Customer Breakthrough',
+    url_path: '/customer-breakthroughs',
+  },
+  'acquisition-and-analysis-software': {
+    name: 'Acquisition and Analysis Software',
+  },
+  'igg-quantification-assays': {
+    name: 'IgG quantitation',
+  },
+  'dna-quantitation': {
+    name: 'DNA Quantitation',
+  },
+  elisa: {
+    name: 'ELISA',
+  },
+  'cell-viability': {
+    name: 'Cell Viability',
+  },
+  cardiotox: {
+    name: 'Cardiotox',
+  },
+  gpcrs: {
+    name: 'GPCR',
+  },
+  'ion-channel': {
+    name: 'Ion Channel',
+  },
+  'reporter-gene-assays': {
+    name: 'Reporter Gene',
+  },
+  'western-blot': {
+    name: 'Western Blot',
+  },
+  transporters: {
+    name: 'Transporter',
+  },
+  'mammalian-screening': {
+    name: 'Mammalian Screening',
+    url_path: '/products/clone-screening/mammalian-screening',
+  },
+  'microbial-screening': {
+    name: 'Microbial Screening',
+    url_path: '/products/clone-screening/microbial-screening',
+  },
+  stacker: {
+    name: 'Stacker',
+  },
+  'single-cell-isolation': {
+    name: 'Single-Cell Isolation',
+  },
+  digitizers: {
+    name: 'Digitizers',
+  },
+  amplifiers: {
+    name: 'Amplifiers',
+  },
+  resources: customResourcesBreadcrumb,
+  events: {
+    name: 'Events',
+    url_path: '/events',
+  },
+  newsletters: {
+    name: 'Newsletters',
+    url_path: '/newsletters',
+  },
+  brochures: customResourcesBreadcrumb,
+  'for-whats-next': {
+    name: 'For What\'s Next',
+    url_path: '/for-whats-next',
+  },
+};
+
+export default customBreadcrumbs;
